Attempting to fix index.js for package load
Is this being compiled as ECMAScript? Then it should have `import` which I had to change. 

Is it React Native? Then it should have ?OptionalTypes which I had to delete.

It still doesn't work regardless of the fact nuclide-commons-atom/FeatureLoader.js has an export default class FeatureLoader but that would only work in ES4/5/6....